requirements.txt
# Web Scraping
scrapy>=2.11.0
scrapy-playwright>=0.0.34
playwright>=1.40.0

# Database
redis>=5.0.1
falkordb>=1.0.6

# AI/ML
google-generativeai>=0.3.2

# Data Validation & Serialization
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Configuration
python-dotenv>=1.0.0

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-asyncio>=0.21.1
pytest-mock>=3.12.0

# Code Quality
black>=23.12.0
pylint>=3.0.3
mypy>=1.7.1
flake8>=6.1.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
requests>=2.31.0
aiohttp>=3.9.1

# Logging
loguru>=0.7.2

# Type hints for libraries
types-redis>=4.6.0
types-requests>=2.31.0
types-python-dateutil>=2.8.19
tqdm==4.67.1
